backend: consolidated the tests
