--- backend/src/api/gameplay/gameplay.start.router.test.ts
@@ -1,92 +1,0 @@
-import { describe, expect, it } from "vitest"
-import { createNewAccountModelStub } from "#api/accounts/NewAccountModel.stub.ts"
-import { createApiStub } from "#api/createApi.stub.ts"
-import { type CreateLobbyDto, type LobbyPlayerDto } from "#api/lobbies/lobbies.controller.ts"
-import { GameStatus } from "#api/shared/GameStatus.ts"
-import { createDefaultStarSystemGenerationSettings } from "#lib/star-systems/createDefaultStarSystemGenerationSettings.ts"
-import { extractSuccess } from "#tests/extractSuccess.ts"
-import { TrpcClient } from "#tests/TrpcClient.ts"
-
-describe("gameplay.router", () => {
-  describe("start", () => {
-    it("should start a game", async () => {
-      // Arrange
-      const { api, authService, accountsRepository } = await createApiStub()
-      using trpcClient = new TrpcClient({ api })
-
-      const creatorAccount = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub()))
-      authService.account = creatorAccount
-
-      const newGameSettings: CreateLobbyDto["configuration"] = {
-        name: "start game",
-        nbSeats: 2,
-        tickIntervalSeconds: 60,
-      }
-
-      const { createdGameId } = await trpcClient.client.lobbies.create.mutate({ configuration: newGameSettings })
-
-      // Act
-      const startGameResult = await trpcClient.client.gameplay.start.mutate({ gameId: createdGameId })
-
-      // Assert
-
-      const creator: LobbyPlayerDto = { id: creatorAccount.id, alias: creatorAccount.alias }
-      expect(startGameResult).toEqual<typeof startGameResult>({
-        id: createdGameId,
-        createdAt: expect.any(String),
-        endedAt: null,
-        winnerAccountId: null,
-        configuration: {
-          ...newGameSettings,
-          starSystemGenerationSettings: {
-            ...createDefaultStarSystemGenerationSettings(),
-            seed: expect.any(Number),
-          },
-        },
-        startedAt: expect.any(String),
-        creator,
-        players: [creator],
-        status: GameStatus.STARTED,
-        canJoin: false, // Because started
-        canLeave: false, // Because started
-        canStart: false, // Because started
-      })
-    })
-
-    it("should reject starting a game as a non-creator", async () => {
-      // Arrange
-      const { api, authService, accountsRepository } = await createApiStub()
-      using trpcClient = new TrpcClient({ api })
-
-      const creator = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub({ alias: "Creator" })))
-      const account = extractSuccess(await accountsRepository.createAccount(createNewAccountModelStub({ alias: "Player 2" })))
-      authService.account = creator
-
-      const { createdGameId } = await trpcClient.client.lobbies.create.mutate({
-        configuration: {
-          name: "non creator cannot start game",
-          nbSeats: 2,
-          tickIntervalSeconds: 60,
-        },
-      })
-
-      authService.account = account
-
-      // Act & Assert
-      await expect(trpcClient.client.gameplay.start.mutate({ gameId: createdGameId })).rejects.toMatchObject({
-        data: { code: "BAD_REQUEST" },
-      })
-    })
-
-    it("should reject anonymous game start", async () => {
-      // Arrange
-      const { api } = await createApiStub()
-      using trpcClient = new TrpcClient({ api })
-
-      // Act & Assert
-      await expect(trpcClient.client.gameplay.start.mutate({ gameId: 1 })).rejects.toMatchObject({
-        data: { code: "UNAUTHORIZED" },
-      })
-    })
-  })
-})
