+++ backend/src/api/listings/listings.controller.ts
@@ -1,6 +1,6 @@
 import { type Logger, Result } from "@guillaume-docquier/tools-ts"
 import { z } from "zod"
-import { GameId } from "#api/games/GameId.ts"
+import { GameId } from "#api/shared/GameId.ts"
 import { GameStatus } from "#api/shared/GameStatus.ts"
 import type { ListingsRepository } from "./listings.repository.ts"
