+++ backend/src/api/lobbies/lobbies.controller.ts
@@ -1,9 +1,9 @@
 import { type Logger, Result } from "@guillaume-docquier/tools-ts"
 import z from "zod"
 import { AccountId } from "#api/accounts/AccountId.ts"
-import { GameId } from "#api/games/GameId.ts"
-import { PlayerId } from "#api/games/PlayerId.ts"
+import { GameId } from "#api/shared/GameId.ts"
 import { computeGameStatus, GameStatus } from "#api/shared/GameStatus.ts"
+import { PlayerId } from "#api/shared/PlayerId.ts"
 import { RangeDto } from "#api/shared/RangeDto.ts"
 import type { CreateTransaction } from "#lib/db/createDb.ts"
 import { couldNot, rollbackOnFailure, TransactionRollback } from "#lib/errors.ts"
