+++ backend/src/api/star-systems/starSystems.controller.ts
@@ -1,8 +1,8 @@
 import { type Failure, type Logger, Result } from "@guillaume-docquier/tools-ts"
 import { z } from "zod"
 import type { AccountId } from "#api/accounts/AccountId.ts"
-import type { GameId } from "#api/games/GameId.ts"
 import { type LobbiesRepository } from "#api/lobbies/lobbies.repository.ts"
+import type { GameId } from "#api/shared/GameId.ts"
 import { RangeDto } from "#api/shared/RangeDto.ts"
 import type { StarSystemsRepository } from "#lib/db/star-systems/starSystems.repository.ts"
 import { notAuthorized } from "#lib/errors.ts"
