+++ backend/src/api/star-systems/starSystems.router.test.ts
@@ -3,8 +3,8 @@ import { v4 } from "uuid"
 import { describe, expect, it } from "vitest"
 import { createNewAccountModelStub } from "#api/accounts/NewAccountModel.stub.ts"
 import { createApiStub } from "#api/createApi.stub.ts"
-import type { GameId } from "#api/games/GameId.ts"
 import { createCreateLobbyDtoStub } from "#api/lobbies/CreateLobbyDto.stub.ts"
+import type { GameId } from "#api/shared/GameId.ts"
 import { type NewStarSystemModel, type StarSystemsRepository } from "#lib/db/star-systems/starSystems.repository.ts"
 import { BodyType } from "#lib/star-systems/BodyType.ts"
 import { extractSuccess } from "#tests/extractSuccess.ts"
