+++ backend/src/lib/db/resources/gamePlayerResources.repository.ts
@@ -1,7 +1,7 @@
 import { type Logger, Result } from "@guillaume-docquier/tools-ts"
 import { and, eq, sql } from "drizzle-orm"
-import type { GameId } from "#api/games/GameId.ts"
-import type { PlayerId } from "#api/games/PlayerId.ts"
+import type { GameId } from "#api/shared/GameId.ts"
+import type { PlayerId } from "#api/shared/PlayerId.ts"
 import { PostgresRepository } from "#lib/db/PostgresRepository.ts"
 import { gamePlayerResourcesTable } from "#lib/db/schema.ts"
 import { couldNot } from "#lib/errors.ts"
