+++ backend/src/lib/db/star-systems/starSystems.repository.test.ts
@@ -4,10 +4,10 @@ import { v4 } from "uuid"
 import { describe, expect, it } from "vitest"
 import { AccountsRepository } from "#api/accounts/accounts.repository.ts"
 import { createNewAccountModelStub } from "#api/accounts/NewAccountModel.stub.ts"
-import type { GameId } from "#api/games/GameId.ts"
 import { createCreateLobbyDtoStub } from "#api/lobbies/CreateLobbyDto.stub.ts"
 import { LobbiesController } from "#api/lobbies/lobbies.controller.ts"
 import { LobbiesRepository } from "#api/lobbies/lobbies.repository.ts"
+import type { GameId } from "#api/shared/GameId.ts"
 import { createDbMock } from "#lib/db/createDb.mock.ts"
 import type { Database } from "#lib/db/createDb.ts"
 import { bodiesTable, movementEdgesTable, movementNodesTable, orbitsTable, sectorsTable, starSystemsTable } from "#lib/db/schema.ts"
